+++ tests/test_add.py
@@ -151,26 +151,6 @@ def test_determine_remote_returns_none_for_empty_remotes():
     assert result is None
 
 
-def test_determine_remote_requires_path_boundary():
-    """An org-scoped remote must not match a different org sharing its prefix."""
-    m = Mock()
-    m.remotes = [_make_remote("myorg", "https://github.com/myorg")]
-    result = Manifest.find_remote_for_url(
-        m, "https://github.com/myorg-private/repo.git"
-    )
-    assert result is None
-
-
-def test_determine_remote_matches_exact_and_subpath():
-    """The boundary check still matches the remote itself and any URL beneath it."""
-    m = Mock()
-    m.remotes = [_make_remote("myorg", "https://github.com/myorg")]
-    assert Manifest.find_remote_for_url(m, "https://github.com/myorg") is not None
-    assert (
-        Manifest.find_remote_for_url(m, "https://github.com/myorg/repo.git") is not None
-    )
-
-
 # ---------------------------------------------------------------------------
 # Add command – non-interactive
 # ---------------------------------------------------------------------------
